requirements.txt
python-telegram-bot==13.1
flask==1.1.4
requests==2.27.1
python-dotenv==0.19.2
markupsafe==2.0.1
